requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl-cffi==0.11.1
edge-tts==7.0.2
faiss-cpu==1.11.0
fastapi==0.115.12
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
gitpython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-core==0.3.62
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.42
llvmlite==0.44.0
markupsafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
multitasking==0.0.11
narwhals==1.41.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper==20240930
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyaudio==0.2.14
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydeck==0.9.1
pyobjc==11.0
pyobjc-core==11.0
pyobjc-framework-accessibility==11.0
pyobjc-framework-accounts==11.0
pyobjc-framework-addressbook==11.0
pyobjc-framework-adservices==11.0
pyobjc-framework-adsupport==11.0
pyobjc-framework-applescriptkit==11.0
pyobjc-framework-applescriptobjc==11.0
pyobjc-framework-applicationservices==11.0
pyobjc-framework-apptrackingtransparency==11.0
pyobjc-framework-audiovideobridging==11.0
pyobjc-framework-authenticationservices==11.0
pyobjc-framework-automaticassessmentconfiguration==11.0
pyobjc-framework-automator==11.0
pyobjc-framework-avfoundation==11.0
pyobjc-framework-avkit==11.0
pyobjc-framework-avrouting==11.0
pyobjc-framework-backgroundassets==11.0
pyobjc-framework-browserenginekit==11.0
pyobjc-framework-businesschat==11.0
pyobjc-framework-callkit==11.0
pyobjc-framework-carbon==11.0
pyobjc-framework-cfnetwork==11.0
pyobjc-framework-cinematic==11.0
pyobjc-framework-classkit==11.0
pyobjc-framework-cloudkit==11.0
pyobjc-framework-cocoa==11.0
pyobjc-framework-colorsync==11.0
pyobjc-framework-contacts==11.0
pyobjc-framework-contactsui==11.0
pyobjc-framework-coreaudio==11.0
pyobjc-framework-coreaudiokit==11.0
pyobjc-framework-corebluetooth==11.0
pyobjc-framework-coredata==11.0
pyobjc-framework-corehaptics==11.0
pyobjc-framework-corelocation==11.0
pyobjc-framework-coremedia==11.0
pyobjc-framework-coremediaio==11.0
pyobjc-framework-coremidi==11.0
pyobjc-framework-coreml==11.0
pyobjc-framework-coremotion==11.0
pyobjc-framework-coreservices==11.0
pyobjc-framework-corespotlight==11.0
pyobjc-framework-coretext==11.0
pyobjc-framework-corewlan==11.0
pyobjc-framework-cryptotokenkit==11.0
pyobjc-framework-datadetection==11.0
pyobjc-framework-devicecheck==11.0
pyobjc-framework-devicediscoveryextension==11.0
pyobjc-framework-discrecording==11.0
pyobjc-framework-discrecordingui==11.0
pyobjc-framework-diskarbitration==11.0
pyobjc-framework-dvdplayback==11.0
pyobjc-framework-eventkit==11.0
pyobjc-framework-exceptionhandling==11.0
pyobjc-framework-executionpolicy==11.0
pyobjc-framework-extensionkit==11.0
pyobjc-framework-externalaccessory==11.0
pyobjc-framework-fileprovider==11.0
pyobjc-framework-fileproviderui==11.0
pyobjc-framework-findersync==11.0
pyobjc-framework-fsevents==11.0
pyobjc-framework-gamecenter==11.0
pyobjc-framework-gamecontroller==11.0
pyobjc-framework-gamekit==11.0
pyobjc-framework-gameplaykit==11.0
pyobjc-framework-healthkit==11.0
pyobjc-framework-imagecapturecore==11.0
pyobjc-framework-installerplugins==11.0
pyobjc-framework-intents==11.0
pyobjc-framework-intentsui==11.0
pyobjc-framework-iobluetooth==11.0
pyobjc-framework-iobluetoothui==11.0
pyobjc-framework-iosurface==11.0
pyobjc-framework-ituneslibrary==11.0
pyobjc-framework-kernelmanagement==11.0
pyobjc-framework-latentsemanticmapping==11.0
pyobjc-framework-launchservices==11.0
pyobjc-framework-libdispatch==11.0
pyobjc-framework-libxpc==11.0
pyobjc-framework-linkpresentation==11.0
pyobjc-framework-localauthentication==11.0
pyobjc-framework-localauthenticationembeddedui==11.0
pyobjc-framework-mailkit==11.0
pyobjc-framework-mapkit==11.0
pyobjc-framework-mediaaccessibility==11.0
pyobjc-framework-mediaextension==11.0
pyobjc-framework-medialibrary==11.0
pyobjc-framework-mediaplayer==11.0
pyobjc-framework-mediatoolbox==11.0
pyobjc-framework-metal==11.0
pyobjc-framework-metalfx==11.0
pyobjc-framework-metalkit==11.0
pyobjc-framework-metalperformanceshaders==11.0
pyobjc-framework-metalperformanceshadersgraph==11.0
pyobjc-framework-metrickit==11.0
pyobjc-framework-mlcompute==11.0
pyobjc-framework-modelio==11.0
pyobjc-framework-multipeerconnectivity==11.0
pyobjc-framework-naturallanguage==11.0
pyobjc-framework-netfs==11.0
pyobjc-framework-network==11.0
pyobjc-framework-networkextension==11.0
pyobjc-framework-notificationcenter==11.0
pyobjc-framework-opendirectory==11.0
pyobjc-framework-osakit==11.0
pyobjc-framework-oslog==11.0
pyobjc-framework-passkit==11.0
pyobjc-framework-pencilkit==11.0
pyobjc-framework-phase==11.0
pyobjc-framework-photos==11.0
pyobjc-framework-photosui==11.0
pyobjc-framework-preferencepanes==11.0
pyobjc-framework-pushkit==11.0
pyobjc-framework-quartz==11.0
pyobjc-framework-quicklookthumbnailing==11.0
pyobjc-framework-replaykit==11.0
pyobjc-framework-safariservices==11.0
pyobjc-framework-safetykit==11.0
pyobjc-framework-scenekit==11.0
pyobjc-framework-screencapturekit==11.0
pyobjc-framework-screensaver==11.0
pyobjc-framework-screentime==11.0
pyobjc-framework-searchkit==11.0
pyobjc-framework-security==11.0
pyobjc-framework-securityfoundation==11.0
pyobjc-framework-securityinterface==11.0
pyobjc-framework-sensitivecontentanalysis==11.0
pyobjc-framework-servicemanagement==11.0
pyobjc-framework-sharedwithyou==11.0
pyobjc-framework-sharedwithyoucore==11.0
pyobjc-framework-shazamkit==11.0
pyobjc-framework-social==11.0
pyobjc-framework-soundanalysis==11.0
pyobjc-framework-speech==11.0
pyobjc-framework-spritekit==11.0
pyobjc-framework-storekit==11.0
pyobjc-framework-symbols==11.0
pyobjc-framework-syncservices==11.0
pyobjc-framework-systemconfiguration==11.0
pyobjc-framework-systemextensions==11.0
pyobjc-framework-threadnetwork==11.0
pyobjc-framework-uniformtypeidentifiers==11.0
pyobjc-framework-usernotifications==11.0
pyobjc-framework-usernotificationsui==11.0
pyobjc-framework-videosubscriberaccount==11.0
pyobjc-framework-videotoolbox==11.0
pyobjc-framework-virtualization==11.0
pyobjc-framework-vision==11.0
pyobjc-framework-webkit==11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
speechrecognition==3.14.3
sqlalchemy==2.0.41
srt==3.5.3
starlette==0.46.2
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.3
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
unicorn==2.1.3
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
yfinance==0.2.61
zstandard==0.23.0
